Split tangled types.ts file into domain files
